Created caesarCipher.c
This is still a work in progress, but the functionality is more or less there.  

Implements a Caesar cipher, both encrypting and decrypting. The key is an integer. This cipher performs its encryption by replacing 
each character in the message by 'shifting' the ASCII value of the character up by a given number of places (the key). i.e. key 2 
encrypts "HI" to "JK".